% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{knit_print.gog_page}
\alias{knit_print.gog_page}
\title{knitr print method for a composed page.}
\usage{
knit_print.gog_page(x, ...)
}
\arguments{
\item{x}{The plot or page being displayed.}

\item{...}{Passed on by the display host; unused.}
}
\description{
knitr print method for a composed page.
}
