% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_svg}
\alias{render_svg}
\title{Render a gog_spec to an SVG string.}
\usage{
render_svg(gog)
}
\arguments{
\item{gog}{A \code{gog_spec} object built with \code{data() + ...}.}
}
\value{
A character string containing the SVG.
}
\description{
Render a gog_spec to an SVG string.
}
