% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{repr_html.gog_page}
\alias{repr_html.gog_page}
\title{repr method: a composed page as inline SVG in a Jupyter cell.}
\usage{
repr_html.gog_page(obj, ...)
}
\arguments{
\item{obj}{The plot or page being displayed.}

\item{...}{Passed on by the display host; unused.}
}
\description{
repr method: a composed page as inline SVG in a Jupyter cell.
}
