% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{repr_text.gog_page}
\alias{repr_text.gog_page}
\title{repr method: a one-line description of a composed page.}
\usage{
repr_text.gog_page(obj, ...)
}
\arguments{
\item{obj}{The plot or page being displayed.}

\item{...}{Passed on by the display host; unused.}
}
\description{
repr method: a one-line description of a composed page.
}
