% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{*.gog_atom}
\alias{*.gog_atom}
\title{Derive a compound layer atom from a mark and a transform.}
\usage{
\method{*}{gog_atom}(e1, e2)
}
\arguments{
\item{e1}{The mark, or a layer already carrying one transform.}

\item{e2}{The transform being joined to it.}
}
\value{
A layer \code{gog_atom} carrying the mark and its transforms.
}
\description{
Derive a compound layer atom from a mark and a transform.
}
